Update README.md
3358 Data Structures and Algorithms - Dr.Koh
Assignment Description:
You will use a fixed-sized, compile-time array (and other supporting data members) to implement an IntSet data type (using C++ class with member variables declared private) that can be used to declare variables (objects) each of which can represent a set of up to MAX_SIZE integers. For the purpose of testing, set MAX_SIZE to 10; your design and implementation of the data type, however, should enable the maximum size to be easily modified, i.e., only need to change the value that MAX_SIZE is set to.